+++ react-app/src/pages/LoginPage/VolunteerLoginPage/VolunteerLoginPage.module.css
@@ -1,0 +1,110 @@
+/* split screen in two */
+.split {
+    height: 100%;
+    position: fixed;
+    z-index: 1;
+    top: 0;
+    overflow-x: hidden;
+}
+
+/* left image panel */
+.left {
+    left: 0;
+    width: 65%;
+    max-width: calc(100vw - 600px);
+}
+
+/* right login panel */
+.right {
+    right: 0;
+    width: 35%;
+    min-width: 600px;
+}
+
+/* center right login panel vertically & horizontally */
+.centered {
+    position: absolute;
+    top: 50%;
+    left: 50%;
+    transform: translate(-50%, -50%);
+    text-align: center;
+  }
+
+/* login fill in section */
+.login_input{
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    width: 100%;
+    height: 100%;
+}
+
+/* welcome! */
+.heading{
+    font-size: 33px;
+    color: var(--blue-gray);
+    /* margin: 8px;
+    margin-top: 5%; */
+    margin-bottom: 0%;
+}
+
+/* login text */
+.alignLeft{
+    width: 100%;
+    text-align: left;
+}
+
+/* for input text boxes */
+.label{
+    margin-bottom: 3%; 
+    margin-top: 8%;
+    color: var(--blue-gray);
+}
+
+/* switch to admin login - style to look like link */
+.switch{
+    font-weight: bold;
+    text-decoration: underline;
+    text-underline-position: under;
+    background-color: white;
+    border: none;
+    cursor: pointer;
+    margin: 8% 0% 5% 0%
+}
+
+/* for left panel */
+.leftImgContainer {
+    width: 100%;
+    height: 100%;
+    overflow: hidden;
+}
+
+.leftImgContainer img {
+    width: 100%;
+    height: 100%;
+    object-fit: cover;
+}
+
+/* for logo */
+.rightImgContainer {
+    width: 43%;
+    height: 43%;
+    overflow: hidden;
+}
+
+.rightImgContainer img {
+    width: 100%;
+    height: 100%;
+    object-fit: cover;
+}
+
+/* mobile */
+@media (max-width: 600px) {
+    .left {
+        width: 0%;
+    }
+    .right {
+        width: 100%;
+        min-width: 0%;
+    }
+}
